requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
python-dotenv==1.0.0
azure-storage-blob==12.24.0
azure-search-documents==11.5.3
azure-identity==1.19.0
azure-core==1.32.0
azure-ai-documentintelligence>=1.0.0b3
beautifulsoup4>=4.12.3
openai==1.109.1
supabase==2.10.0
pydantic>=2.10.4
PyPDF2==3.0.1
openpyxl==3.1.5
pandas==2.2.3
